In memory implementation of ink link backend
Particularly proud of the player order permutations. To guarantee that
each player of N players gets to see every book, gets an equal number of
draws and describes *and* isn't always after the same player in a round
we need N permutations of the player orders which share no common slots.
This turns out to be quite expensive to compute so we pre-compute the
order selections and shuffle the player numbers on each round.

Trying to compute all the permutations of 11 players caused node to run
out of memory and I could not be bothered to figure out how to give it
more so if you have 11+ players we just pick an order and then cycle
through N times (this simulates sitting in a circle and passing the
books the same way each time) order will still be random between rounds.