Remove mutators on beans
Pre-cursor to doing this all in a DB, we just want app to do reads on
beans it gets back from the services. Likewise the services should be
doing the work to update the state.

Unwind the entanglement between JobService and BookService, instead we
have the app make the right calls on both services.