Refactor disparate Services into Persistence
Pull all the Service interfaces and in memory implementations into a
package.

Wrap the services up into one interface with fields for the sub services
and then wrap that in a Persistence.

Persistence takes:
 - a service provider
 - a function for read only access to the persistence
 - a function to start a transaction
 (which both return the T for the service provider)
 - a function to commit the transaction
 - a function to abort the transaction

In exchange you get a PersistenceProvider which can give you a Promise
for a Services for reading or take a function that produces your update
and give you a promise for that update.

Incredibly this all just worked with an in memory service provider.
Tomorrow's challenge is writing a db service provider.🤞