Fix a few logic bugs
Delete a book if we're skipping the first page.

Don't clear active gallery id if you didn't know which gallery id you
were trying to clear.

Actually skip the page in the inactive job poller.

Invoke the complete function instead of checking it exists. Surprised
typescript did not warn me about that one but perhaps it is because the
book may be undefined in that block.